Update on "Give type for joint_custom_pass"
[ghstack-poisoned]
